+++ CHANGELOG.md
@@ -7,14 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fix(validation): Show correct max counterparty collateral in validation message.
 - Fix(mobile): Calculate counterparty balance correctly when checking the validity of trade parameters.
+- Fix(trade): Spawn dedicated tokio task when executing trade
 
 ## [1.8.8] - 2024-02-14
 
 - Feat(mobile): Let user add a name to their profile for the leaderboard
 - Feat(coordinator): Allow to specify time range for leadership board
 - Fix: Prevent trade failures by ensuring that the funding transaction is not overpaying fees unexpectedly.
-- Fix(validation): Show correct max counterparty collateral in validation message.
 
 ## [1.8.7] - 2024-02-10